requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
Django==5.0.4
psycopg2-binary==2.9.9

# ── HTTP ─────────────────────────────────────────────────────────────────────
requests==2.31.0
urllib3==2.2.1

# ── Data processing ───────────────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
pyarrow==16.0.0        # parquet support

# ── Config ────────────────────────────────────────────────────────────────────
PyYAML==6.0.1
